backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
xgboost==2.1.3
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
joblib==1.4.2
stellar-sdk==10.0.0
httpx==0.27.0
python-multipart==0.0.9
aiosqlite==0.20.0
